tools/cleanup.sh: make it work with newer opam
--with-test behaves in a different way now: it only applies to packages you
explicitly list, and not their dependencies.

So first need to find out all the dependencies of xs-toolstack (that'll be all
our xs packages), and their recursive dependencies.
And we want the test dependencies of all of these (but not the test
dependencies of test-only packages, because that might pull in too many
packages and often it is not even possible to solve this with a consistent
version of packages, e.g. different packages may have conflicting constraints
on their test dependencies).

Also add some additional consistency check with another 'opam admin' command
that ignores doc and test dependencies (doc dependencies introduce cycles...)
that can help when updating opam packages.

Currently this prints an error that 'mdx' is not coinstallable, but that can be
solved by updating 'mdx'.

For debug purposes add a more direct installtest.sh that will install all these
packages.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>